refactor(github-projects): address review — extract chip color module, localize no-value column label
- Move ChipColors/chipStyle/singleSelectChipColors/labelChipColors from
  ProjectCell.tsx into single-select-chip-colors.ts so the table cell
  component no longer exports rendering utilities for the board to import.
- Localize the board's trailing "No <field>" column label via translate()
  in the renderer (the shared column builder keeps the English fallback,
  since translate() is renderer-only), with ko/ja/zh/es catalog entries.
