+++ src/renderer/src/i18n/locales/es.json
@@ -2161,7 +2161,8 @@
           "ProjectBoardView": {
             "no-group-field": "Este tablero no tiene un campo de selección única para agrupar.",
             "empty": "Ningún elemento coincide con el filtro de esta vista.",
-            "restricted": "Elemento restringido"
+            "restricted": "Elemento restringido",
+            "no-value-column": "Sin {{value0}}"
           }
         },
         "GitHubMarkdownComposer": {
