+++ src/renderer/src/i18n/locales/ko.json
@@ -2161,7 +2161,8 @@
           "ProjectBoardView": {
             "no-group-field": "이 보드에는 그룹화할 단일 선택 필드가 없습니다.",
             "empty": "이 보기의 필터와 일치하는 항목이 없습니다.",
-            "restricted": "제한된 항목"
+            "restricted": "제한된 항목",
+            "no-value-column": "{{value0}} 없음"
           }
         },
         "GitHubMarkdownComposer": {
